Index from judgements created now you can select a problem from the problems window or specificly the problem and try to submit any problem then the EvaluateTool evaluete file size, time limit, memory limit with the values of the problem

+++ database/migrations/2016_09_20_004744_create_judgments_table.php
@@ -18,7 +18,7 @@ class CreateJudgmentsTable extends Migration
             $table->dateTime('submitted_at');
             $table->string('language',8);
             $table->integer('memory');
-            $table->integer('time');
+            $table->float('time');
             $table->string('judgment',40); // the submit´s verdict
             $table->integer('file_size');
             $table->integer('problem_id')->unsigned();
